celery-requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
flower==2.0.1
humanize==4.9.0
kombu==5.3.4
loguru==0.7.2
prometheus-client==0.19.0
prompt-toolkit==3.0.41
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
vine==5.1.0
wcwidth==0.2.12
